Raise error if device type is not found (#12)
* Raise error if device type is not found

The Setpoint 161 has been observed sending an invalid device type if it powers on after or at the same time as the Gateway 482. Power cycling the gateway after the 161 is powered on seems to fix it, but the integration can't proceed with the bad type.

* Device may reappear after start up

The 161 may reappear on the network after integration startup, but it's not in inventory.